spanner: asynchronous session pool init
Initialize the session pool asynchronously when creating the spanner
client instead of blocking on session creation.

This matches the golang client behavior and is useful in applications
where application startup latency is important. Although such
applications can themselves asynchronously instantiate the spanner
client, this adds complexity to the applications. As the spanner client
already has the infrastructure in place to make this initialization
asynchronous it seems reasonable to lift it into the spanner client.